restructure methods
